+++ tools/mailq.cpp
@@ -45,7 +45,7 @@ int main()
 		strftime(timebuf, sizeof(timebuf), "%FT%T", localtime(&sb.st_mtime));
 		printf("%-6s  %-19s", de->d_name, timebuf);
 		size_t clump_size = sizeof(size_t) + 3 * sizeof(uint32_t) + 2;
-		if (static_cast<size_t>(sb.st_size) < clump_size) {
+		if (static_cast<unsigned long long>(sb.st_size) < clump_size) {
 			printf("\n");
 			continue;
 		}
